Added Basics File
Added the new Basic Usage fie which is a slimmed down
and reworked version of the current Usage file. The idea is that the
minimal information needed to use Stash will be on this page, so people
who just need to use it (not integrate into a framework or extend it,
but use it as part of an existing framework or project) can dive right
in.